feat: integrate UploadThing for file handling and enhance user experience with new components
