+++ apps/app/app/api/uploadthing/delete/route.ts
@@ -1,0 +1,43 @@
+import { auth } from '@repo/auth/server';
+import { headers } from 'next/headers';
+import { NextResponse } from 'next/server';
+import { UTApi } from 'uploadthing/server';
+
+const utapi = new UTApi();
+
+export async function POST(request: Request) {
+  try {
+    // Authenticate the request using the correct headers approach
+    const session = await auth.api.getSession({
+      headers: await headers(),
+    });
+
+    if (!session?.user) {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
+    // Get file key from request body
+    const { fileKey } = await request.json();
+    
+    if (!fileKey) {
+      return NextResponse.json(
+        { error: 'File key is required' },
+        { status: 400 }
+      );
+    }
+
+    // Delete the file using UTApi
+    await utapi.deleteFiles(fileKey);
+
+    return NextResponse.json(
+      { success: true, message: 'File deleted successfully' },
+      { status: 200 }
+    );
+  } catch (error) {
+    console.error('Error deleting file:', error);
+    return NextResponse.json(
+      { error: 'Failed to delete file' },
+      { status: 500 }
+    );
+  }
+}
