+++ apps/app/app/api/uploadthing/route.ts
@@ -1,0 +1,4 @@
+import { router } from '@/utils/upload';
+import { createRouteHandler } from '@repo/storage';
+
+export const { GET, POST } = createRouteHandler({ router });
