+++ apps/app/app/layout.tsx
@@ -3,6 +3,9 @@ import { DesignSystemProvider } from '@repo/design-system';
 import { fonts } from '@repo/design-system/lib/fonts';
 import { Toolbar } from '@repo/feature-flags/components/toolbar';
 import type { ReactNode } from 'react';
+import { extractRouterConfig } from '@repo/storage';
+import { StorageSSRPlugin } from '@repo/storage/ssr';
+import { router } from '@/utils/upload';
 
 type RootLayoutProperties = {
   readonly children: ReactNode;
@@ -14,6 +17,8 @@ const RootLayout = ({ children }: RootLayoutProperties) => (
       <meta name="apple-mobile-web-app-title" content="Formflow" />
     </head>
     <body>
+      <StorageSSRPlugin routerConfig={extractRouterConfig(router)} />
+
       <DesignSystemProvider>{children}</DesignSystemProvider>
       <Toolbar />
     </body>
