+++ apps/app/package.json
@@ -49,7 +49,7 @@
     "pako": "^2.1.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-hook-form": "^7.54.2",
+    "react-hook-form": "^7.56.2",
     "react-resizable-panels": "^2.1.7",
     "react-wrap-balancer": "^1.1.1",
     "require-in-the-middle": "^7.4.0",
