+++ apps/app/utils/uploadthing-helper.ts
@@ -1,0 +1,5 @@
+import { generateReactHelpers } from "@repo/storage/client";
+
+import type { router } from "@/utils/upload";
+
+export const { useUploadThing } = generateReactHelpers<typeof router>();
