+++ package.json
@@ -43,7 +43,9 @@
     "@clack/prompts": "^0.9.1",
     "@tanstack/react-query": "^5.67.1",
     "commander": "^13.1.0",
-    "framer-motion": "^12.9.2"
+    "framer-motion": "^12.9.2",
+    "sonner": "^1.7.2",
+    "uploadthing": "^7.6.0"
   },
   "type": "module"
 }
