+++ packages/design-system/hooks/use-uploadthing.ts
@@ -1,0 +1,5 @@
+import { generateReactHelpers } from "@repo/storage/client";
+
+import type { router } from '@repo/design-system/lib/uploadthing';
+
+export const { useUploadThing } = generateReactHelpers<typeof router>();
