+++ packages/design-system/package.json
@@ -37,7 +37,6 @@
     "@radix-ui/react-toggle-group": "^1.1.1",
     "@radix-ui/react-tooltip": "^1.1.6",
     "@repo/analytics": "workspace:*",
-    "@repo/auth": "workspace:*",
     "@repo/observability": "workspace:*",
     "@repo/schema-types": "workspace:*",
     "@tanstack/react-query-devtools": "^5.67.1",
@@ -51,13 +50,14 @@
     "input-otp": "^1.4.2",
     "lucide-react": "^0.473.0",
     "motion": "^12.0.6",
+    "next": "15.3.1",
     "next-themes": "^0.4.4",
     "nuqs": "^2.3.2",
     "pako": "^2.1.0",
     "react": "19.0.0",
     "react-aria-components": "^1.6.0",
     "react-day-picker": "^8.10.1",
-    "react-hook-form": "^7.54.2",
+    "react-hook-form": "^7.56.2",
     "react-moveable": "^0.56.0",
     "react-phone-number-input": "^3.4.11",
     "react-resizable-panels": "^2.1.7",
