+++ packages/storage/index.ts
@@ -1,1 +1,6 @@
-export * from '@vercel/blob';
+import { createUploadthing } from 'uploadthing/next';
+
+export { type FileRouter, createRouteHandler } from 'uploadthing/next';
+export { UploadThingError as UploadError, extractRouterConfig } from 'uploadthing/server';
+
+export const storage: ReturnType<typeof createUploadthing> = createUploadthing();
