+++ packages/storage/keys.ts
@@ -4,13 +4,9 @@ import { z } from 'zod';
 export const keys = () =>
   createEnv({
     server: {
-      BLOB_READ_WRITE_TOKEN: z.string().min(1).optional(),
-      PROJECT_ID: z.string(),
-      REGION: z.string(),
+      UPLOADTHING_TOKEN: z.string().min(1).optional(),
     },
     runtimeEnv: {
-      BLOB_READ_WRITE_TOKEN: process.env.BLOB_READ_WRITE_TOKEN,
-      PROJECT_ID: process.env.PROJECT_ID,
-      REGION: process.env.REGION,
+      UPLOADTHING_TOKEN: process.env.UPLOADTHING_TOKEN,
     },
   });
