+++ packages/storage/package.json
@@ -8,9 +8,12 @@
   },
   "dependencies": {
     "@t3-oss/env-nextjs": "^0.11.1",
-    "@vercel/blob": "^0.27.1",
+    "@uploadthing/react": "^7.3.0",
     "appwrite": "14.0.1",
     "zod": "^3.24.1"
+  },
+  "peerDependencies": {
+    "uploadthing": "^7.6.0"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*"
