+++ packages/storage/ssr.ts
@@ -1,0 +1,1 @@
+export { NextSSRPlugin as StorageSSRPlugin } from '@uploadthing/react/next-ssr-plugin';
