+++ packages/tailwind-config/package.json
@@ -11,6 +11,9 @@
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7"
   },
+  "peerDependencies": {
+    "uploadthing": "^7.6.0"
+  },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "tailwindcss": "^3.4.17"
